refactior pt1
