+++ src/js/ui/widgets/Popup.js
@@ -2,10 +2,10 @@ import React, { useEffect, useState } from "react";
 import { Table, TableContainer, TableHead, TableCell, TableRow, TableBody, Paper, makeStyles, Drawer, Typography, IconButton, Grid } from "@material-ui/core";
 import { useGlobalState } from "../global/GlobalState";
 import Util from "../../library/apertureUtil";
-import fipsToState from "../../../json/fipsToState.json"
-import defaultImportantFields from "../../../json/defaultImportantFields.json"
 import CloseIcon from "@material-ui/icons/Close";
 import PopupTimeChart from "./PopupTimeChart";
+import { keyToDisplay, valueToDisplay, keyValueIsValid } from "./PopupUtils";
+import defaultImportantFields from "../../../json/defaultImportantFields.json"
 
 const drawerWidth = '450px';
 
@@ -44,82 +44,6 @@ export default function Popup() {
         return () => { window.setPopupObj = () => { } };
     }, [])
 
-    const keyValueIsValid = (key, value) => {
-        if (['meta', 'id', '_id'].includes(key)) {
-            return false;
-        }
-        return true;
-    }
-
-    const keyToDisplay = (key) => {
-        if (obj?.meta?.[key]?.label) {
-            return obj.meta[key].label;
-        }
-        if (defaultImportantFields[key]) {
-            return defaultImportantFields[key].label ?? Util.cleanUpString(key);
-        }
-
-        return Util.cleanUpString(key);
-    }
-
-    const valueToDisplay = (key, value) => {
-        let unit = obj?.meta?.[key]?.unit;
-        if (unit?.toUpperCase() === 'NA') {
-            unit = null;
-        }
-
-        if (obj?.meta?.[key]?.isDate) {
-            return dateToDisplay(value);
-        }
-        else if (defaultImportantFields[key]?.type && !['string', 'number'].includes(defaultImportantFields[key]?.type)) {
-            return specialTypeToDisplay(defaultImportantFields[key].type, value);
-        }
-        else if (['string', 'number'].includes(typeof value)) {
-            return `${value}${unit ? ` ${Util.cleanUpString(unit)}` : ''}`;
-        }
-        else if (typeof value === 'object') {
-            return mongoObjectToSomething(value, (s) => s);
-        }
-        else {
-            return JSON.stringify(value);
-        }
-    }
-
-    const specialTypeToDisplay = (type, value) => {
-        if (type === "stateFips") {
-            return fipsToState[value];
-        }
-    }
-
-    const mongoObjectToSomething = (object, func) => { //this function will be extended as more mongo objects leak in
-        const numericTypes = ['$numberLong', '$numberDecimal'];
-        for (const numericType of numericTypes) {
-            if (object?.[numericType]) {
-                return func(Number(object[numericType]));
-            }
-        }
-        return JSON.stringify(object);
-    }
-
-    const dateToDisplay = (value) => {
-        if (typeof value === 'number') {
-            return epochToDateString(value);
-        }
-        else if (typeof value === 'object') {
-            return mongoObjectToDateString(value);
-        }
-        return JSON.stringify(value)
-    }
-
-    const epochToDateString = (epoch) => {
-        const str = new Date(epoch).toUTCString();
-        return str.substr(0, str.length - 4);
-    }
-
-    const mongoObjectToDateString = (object) => {
-        return mongoObjectToSomething(object, epochToDateString);
-    }
-
     const makeTable = (keyValPairs) => {
         if (!keyValPairs || !keyValPairs.length) {
             return;
@@ -138,9 +62,9 @@ export default function Popup() {
                         .map(([key, value]) => (
                             <TableRow key={key}>
                                 <TableCell component="th" scope="row">
-                                    {keyToDisplay(key)}
+                                    {keyToDisplay(obj, key)}
                                 </TableCell>
-                                <TableCell>{valueToDisplay(key, value)}</TableCell>
+                                <TableCell>{valueToDisplay(obj, key, value)}</TableCell>
                             </TableRow>
                         ))}
                 </TableBody>
@@ -154,7 +78,7 @@ export default function Popup() {
             const importantFields = Object.entries(obj.properties)
                 .filter(([key, value]) => obj.properties?.meta?.[key]?.important || defaultImportantFields[key])
                 .filter(([key, value]) => {
-                    const keyDisp = keyToDisplay(key);
+                    const keyDisp = keyToDisplay(obj, key);
                     if (importantNames.has(keyDisp)) {
                         return false;
                     }
